% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chdir_results.R
\name{chdir_analysis_wrapper}
\alias{chdir_analysis_wrapper}
\title{chdirAnalysis wrapper. Redirects plots to /dev/null and handles data aggregation}
\usage{
chdir_analysis_wrapper(datain, sampleclass, gamma, nnull)
}
\arguments{
\item{datain}{}

\item{sampleclass}{}

\item{nnull}{}

\item{gammas}{}
}
\description{
chdirAnalysis wrapper. Redirects plots to /dev/null and handles data aggregation
}

