% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geode_utils.R
\name{chdir_props}
\alias{chdir_props}
\title{Extract ith element of characteristic
direction properties list from
chdirAnalysis results}
\usage{
chdir_props(chdir, i = 1)
}
\arguments{
\item{chdir}{chdirAnalysis output}

\item{i}{integer}
}
\value{
matrix #{number of genes} * 1 matrix
}
\description{
Extract ith element of characteristic
direction properties list from
chdirAnalysis results
}

