% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geode_utils.R
\name{chdir_results}
\alias{chdir_results}
\title{Extract ith element of the results list from
chdirAnalysis results}
\usage{
chdir_results(chdir, i = 1)
}
\arguments{
\item{chdir}{chdirAnalysis output}

\item{i}{integer}
}
\value{
vector of length <= #{number of genes}
}
\description{
Extract ith element of the results list from
chdirAnalysis results
}

