% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{choose_unique_submissions}
\alias{choose_unique_submissions}
\title{Select ids of the unique submissions}
\usage{
choose_unique_submissions(samples)
}
\arguments{
\item{samples}{tbl_df as returned from preprocess$samples}
}
\value{
list of the unique ids
}
\description{
Select ids of the unique submissions
}

