% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datain.R
\name{datain_log2_transform}
\alias{datain_log2_transform}
\title{log2 transform expression data}
\usage{
datain_log2_transform(datain)
}
\arguments{
\item{datain}{data.frame}
}
\value{
data.frame where columns 2:ncol are log2 transformed
}
\description{
log2 transform expression data
}

