% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datain.R
\name{datain_quantile_normalize}
\alias{datain_quantile_normalize}
\title{quantile normalize expression data}
\usage{
datain_quantile_normalize(datain, add_noise = TRUE)
}
\arguments{
\item{datain}{data.frame}
}
\value{
data.frame where columns 2:ncol are quantile normalized
}
\description{
quantile normalize expression data
}

