% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{disabledActionButton}
\alias{disabledActionButton}
\title{Return disabled button}
\usage{
disabledActionButton(button)
}
\arguments{
\item{shiny::actionButton}{}
}
\value{
shiny::actionButton
}
\description{
Return disabled button
}

