% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{download_data}
\alias{download_data}
\title{Download microtask dataset and}
\usage{
download_data(url = "https://localhost/microtask.csv")
}
\arguments{
\item{url}{character}
}
\value{
data.frame
}
\description{
Download microtask dataset and
}

