\name{example_expression_data}
\alias{example_expression_data}
\docType{data}
\title{
Example Expression Data
}
\description{
This is the data used in the examples.
}
\usage{data(example_expression_data)}
\format{
  The format is:
 chr "example_expression_data"
}


\examples{
data(example_expression_data)
## maybe str(example_expression_data) ; plot(example_expression_data) ...
}
\keyword{datasets}
