\name{example_sampleclass}
\alias{example_sampleclass}
\docType{data}
\title{
Example Sample Class Factor
}
\description{
This is the factor variable that defined the classes of the samples in the examples.
}
\usage{data(example_sampleclass)}
\format{
  The format is:
 chr "example_sampleclass"
}


\examples{
data(example_sampleclass)
## maybe str(example_sampleclass) ; plot(example_sampleclass) ...
}
\keyword{datasets}
