% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{extract_description}
\alias{extract_description}
\title{Extract general information about the dataset}
\usage{
extract_description(dt)
}
\arguments{
\item{dt}{tbl_dt as returned from preprocess_data}
}
\value{
tbl_dt with columns: geo_id, (gene, pert_type | disease_name, disease_id | drug_name, drug_id ),
    organism, cell_type, id
}
\description{
Extract general information about the dataset
}

