% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{extract_samples}
\alias{extract_samples}
\title{Create tidy table with sample ids}
\usage{
extract_samples(dt)
}
\arguments{
\item{dt}{tbl_dt as returned from preprocess_data}
}
\value{
tbl_dt with 3 columns: id, group and sample
}
\description{
Create tidy table with sample ids
}

