% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stats.R
\name{genes_coverage}
\alias{genes_coverage}
\title{Count number of sets covering given gene}
\usage{
genes_coverage(dataset)
}
\arguments{
\item{list}{of dataframes as returned from preprocess}
}
\value{
data.frame with gene and n
}
\description{
Count number of sets covering given gene
}

