% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{list_to_defs}
\alias{list_to_defs}
\title{Convert list of params to list of def tags}
\usage{
list_to_defs(params)
}
\arguments{
\item{params}{list}
}
\value{
container html dl containing (dt, dd) pair for each element of the list
}
\description{
Convert list of params to list of def tags
}

