% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/paea.R
\name{paea_analysis_dispatch_split_both}
\alias{paea_analysis_dispatch_split_both}
\title{Split query and background into up and down and run PAEA on respective parts}
\usage{
paea_analysis_dispatch_split_both(chdirresults, gmtfile, gammas = c(1),
  casesensitive = FALSE)
}
\arguments{
\item{chdirresults}{see GeoDE::chdirAnalysis}

\item{gmtfile}{see GeoDE::chdirAnalysis}

\item{gammas}{see GeoDE::chdirAnalysis}

\item{casesensitive}{see GeoDE::chdirAnalysis}
}
\value{
list with up and down fields
}
\description{
Split query and background into up and down and run PAEA on respective parts
}

