% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/paea.R
\name{paea_analysis_dispatch_split_both_and_reverse}
\alias{paea_analysis_dispatch_split_both_and_reverse}
\title{Split query and background into up and down and run PAEA on  opposite parts}
\usage{
paea_analysis_dispatch_split_both_and_reverse(chdirresults, gmtfile,
  gammas = c(1), casesensitive = FALSE)
}
\arguments{
\item{chdirresults}{see GeoDE::chdirAnalysis}

\item{gmtfile}{see GeoDE::chdirAnalysis}

\item{gammas}{see GeoDE::chdirAnalysis}

\item{casesensitive}{see GeoDE::chdirAnalysis}
}
\value{
list with up (up query vs down background)and down (down query vs up background)
}
\description{
Split query and background into up and down and run PAEA on  opposite parts
}

