% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/paea.R
\name{paea_to_df}
\alias{paea_to_df}
\title{Take PAEA results and return data.frame with
gene set information and pvalue}
\usage{
paea_to_df(paea_results)
}
\arguments{
\item{paea_results}{GeoDE::PAEAAnalysis output}
}
\value{
tbl_df
}
\description{
Take PAEA results and return data.frame with
gene set information and pvalue
}

