% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chdir_results.R
\name{plot_top_genes}
\alias{plot_top_genes}
\title{Plot top genes from chdirAnalysis}
\usage{
plot_top_genes(results)
}
\arguments{
\item{results}{data frame returned from prepare_results}
}
\value{
ggvis plot
}
\description{
Plot top genes from chdirAnalysis
}

