% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chdir_results.R
\name{prepare_down_genes}
\alias{prepare_down_genes}
\title{Extract downregulated genes from results}
\usage{
prepare_down_genes(results)
}
\arguments{
\item{results}{numeric vector extracted from pchdir results}
}
\value{
data.frame
}
\description{
Extract downregulated genes from results
}

