% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{prepare_gene_sets}
\alias{prepare_gene_sets}
\title{Take genes table and prepare list of the gene sets}
\usage{
prepare_gene_sets(genes)
}
\arguments{
\item{genes}{tbl_dt}
}
\value{
list of the character vectors
}
\description{
Take genes table and prepare list of the gene sets
}

