% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/paea.R
\name{prepare_paea_results}
\alias{prepare_paea_results}
\title{Join PAEA results with data description}
\usage{
prepare_paea_results(paea, data_description)
}
\arguments{
\item{paea}{data frame as returned from paea_to_df}

\item{data_description}{tbl_df as returned from extract_description}
}
\value{
tbl_df
}
\description{
Join PAEA results with data description
}

