% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chdir_results.R
\name{prepare_results}
\alias{prepare_results}
\title{Prepare chdir results for plotting}
\usage{
prepare_results(results, n = 40)
}
\arguments{
\item{results}{numeric vector extracted from pchdir results}

\item{n}{numeric max number of genes to keep}
}
\value{
data.frame
}
\description{
Prepare chdir results for plotting
}

