% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chdir_results.R
\name{prepare_up_genes}
\alias{prepare_up_genes}
\title{Extract upregulated genes from results}
\usage{
prepare_up_genes(results)
}
\arguments{
\item{results}{numeric vector extracted from pchdir results}
}
\value{
data.frame
}
\description{
Extract upregulated genes from results
}

