% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Download dataset and extract relevant data}
\usage{
preprocess(url = "https://localhost/microtask.csv")
}
\arguments{
\item{url}{localization of the input data}
}
\value{
list with description, genes and samples dt
}
\description{
Download dataset and extract relevant data
}

