% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chdir.R
\name{preprocess_chdir_input}
\alias{preprocess_chdir_input}
\title{Preprocess input to chdirAnalysis}
\usage{
preprocess_chdir_input(datain)
}
\arguments{
\item{datain}{see GeoDE::chdirAnalysis}
}
\value{
data.frame
}
\description{
Preprocess input to chdirAnalysis
}

