% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{shiny_error_handler}
\alias{shiny_error_handler}
\title{Create error handler function}
\usage{
shiny_error_handler(values)
}
\arguments{
\item{shiny::reactiveValues}{}
}
\value{
function which stores error in values$last_error
}
\description{
Create error handler function
}

