% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{shiny_list_to_li}
\alias{shiny_list_to_li}
\title{Take list with name and href element and return
<li><a href='href'>name</a></li>}
\usage{
shiny_list_to_li(navbar_tab, this_href)
}
\arguments{
\item{navbar_tab}{list}

\item{this_href}{character}
}
\value{
shiny::tags$li
}
\description{
Take list with name and href element and return
<li><a href='href'>name</a></li>
}

