% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{shiny_static_navbar}
\alias{shiny_static_navbar}
\title{Create "static" boostrap navbar}
\usage{
shiny_static_navbar(navbar_brand, navbar_tabs, this_href = "/")
}
\arguments{
\item{navbar_brand}{character page brand}

\item{navbar_tabs}{list of lists each containing name and href}

\item{this_href}{character current url}
}
\value{
shiny::tags$nav
}
\description{
Create "static" boostrap navbar
}

