% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stats.R
\name{submissions_per_cell_type}
\alias{submissions_per_cell_type}
\title{Count number of submissions per cell type}
\usage{
submissions_per_cell_type(dataset)
}
\arguments{
\item{list}{of dataframes as returned from preprocess}
}
\value{
data.frame with cell_type and n
}
\description{
Count number of submissions per cell type
}

