% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datain.R
\name{datain_is_valid}
\alias{datain_is_valid}
\title{Check if datain is a valid es. Experimental.}
\usage{
datain_is_valid(datain)
}
\arguments{
\item{datain}{data.frame}
}
\value{
list with logical field valid, may change in the future
}
\description{
Check if datain is a valid es. Experimental.
}

