% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datain.R
\name{datain_preprocess}
\alias{datain_preprocess}
\title{Apply preprocesing steps to expression data}
\usage{
datain_preprocess(datain, log2_transform = FALSE,
  quantile_normalize = FALSE, id_filter = NULL)
}
\arguments{
\item{datain}{data.frame}

\item{log2_transform}{logical}

\item{quantile_normalize}{logical}
}
\value{
data.table
}
\description{
Apply preprocesing steps to expression data
}

