% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{extract_description}
\alias{extract_description}
\title{Extract general information about the dataset}
\usage{
extract_description(dt)
}
\arguments{
\item{dt}{tbl_dt as returned from preprocess_data}
}
\value{
tbl_dt with columns: id, geo_id, gene, perturbation, species, tissue_cell_line
}
\description{
Extract general information about the dataset
}

