% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{extract_description}
\alias{extract_description}
\title{Extract general information about the dataset}
\usage{
extract_description(df)
}
\arguments{
\item{df}{tbl_df as returned from preprocess_data}
}
\value{
tbl_df with columns: id, geo_accession, gene, perturbation, species, tissue_cell_line
}
\description{
Extract general information about the dataset
}

