% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/paea.R
\name{paea_analysis_dispatch}
\alias{paea_analysis_dispatch}
\title{PAEAAnalysis dispatch function. Should handle separating chdirresults into up and down
and in future some filtering steps}
\usage{
paea_analysis_dispatch(chdirresults, gmtfile, gammas = c(1),
  casesensitive = FALSE, strategy = "up_up+down_down+up_down+down_up",
  with_progress = FALSE)
}
\arguments{
\item{chdirresults}{see GeoDE::PAEAAnalysis}

\item{gmtfile}{see GeoDE::PAEAAnalysis}

\item{gammas}{see GeoDE::PAEAAnalysis}

\item{casesensitive}{see GeoDE::PAEAAnalysis}

\item{strategy}{character}

\item{with_progress}{boolean increment shiny progress bar}
}
\value{
list with paea results
}
\description{
PAEAAnalysis dispatch function. Should handle separating chdirresults into up and down
and in future some filtering steps
}

