% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/paea.R
\name{paea_analysis_wrapper}
\alias{paea_analysis_wrapper}
\title{PAEAAnalysis wrapper. Redirects plots to /dev/null}
\usage{
paea_analysis_wrapper(chdirresults, gmtfile, gammas = c(1),
  casesensitive = FALSE)
}
\arguments{
\item{chdirresults}{see GeoDE::PAEAAnalysis}

\item{gmtfile}{see GeoDE::PAEAAnalysis}

\item{gammas}{see GeoDE::PAEAAnalysis}

\item{casesensitive}{see GeoDE::PAEAAnalysis}
}
\value{
paea results
}
\description{
PAEAAnalysis wrapper. Redirects plots to /dev/null
}

