% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/indata.R
\name{plot_density}
\alias{plot_density}
\title{Create density plots for input data}
\usage{
plot_density(indata)
}
\arguments{
\item{indata}{data.frame with genenames in first columns and samples in following}
}
\value{
ggvis
}
\description{
Create density plots for input data
}

